feat: integrate hightide 0.12.x
