+++ web/components/layout/StagingDisclaimerDialog.tsx
@@ -1,0 +1,58 @@
+'use client'
+
+import { useEffect, useState } from 'react'
+import Link from 'next/link'
+import { Button, Dialog, MarkdownInterpreter } from '@helpwave/hightide'
+import { useStorage } from '@/hooks/useStorage'
+import { getConfig } from '@/utils/config'
+import { useTasksTranslation } from '@/i18n/useTasksTranslation'
+import clsx from 'clsx'
+
+export const StagingDisclaimerDialog = () => {
+  const config = getConfig()
+  const translation = useTasksTranslation()
+
+  const [isStagingDisclaimerOpen, setStagingDisclaimerOpen] = useState(false)
+  const {
+    value: lastTimeStagingDisclaimerDismissed,
+    setValue: setLastTimeStagingDisclaimerDismissed
+  } = useStorage({ key: 'staging-disclaimer-dismissed-time', defaultValue: 0 })
+
+  const dismissStagingDisclaimer = () => {
+    setLastTimeStagingDisclaimerDismissed(new Date().getTime())
+    setStagingDisclaimerOpen(false)
+  }
+
+  useEffect(() => {
+    const ONE_DAY = 1000 * 60 * 60 * 24
+    if (config.showStagingDisclaimerModal && new Date().getTime() - lastTimeStagingDisclaimerDismissed > ONE_DAY) {
+      setStagingDisclaimerOpen(true)
+    }
+  }, [lastTimeStagingDisclaimerDismissed, config.showStagingDisclaimerModal])
+
+  return (
+    <Dialog
+      isModal={false}
+      isOpen={isStagingDisclaimerOpen}
+      onClose={dismissStagingDisclaimer}
+      titleElement={translation('developmentAndPreviewInstance')}
+      description={(<MarkdownInterpreter text={translation('stagingModalDisclaimerMarkdown')} />)}
+      className={clsx('z-20 w-200')}
+      backgroundClassName="z-10"
+    >
+      <div className="flex-row-8">
+        <Link className="text-primary hover:brightness-75 font-bold" href={config.imprintUrl}>
+          {translation('imprint')}
+        </Link>
+        <Link className="text-primary hover:brightness-75 font-bold" href={config.privacyUrl}>
+          {translation('privacy')}
+        </Link>
+      </div>
+      <div className="flex-row-0 justify-end">
+        <Button color="positive" onClick={dismissStagingDisclaimer}>
+          {translation('confirm')}
+        </Button>
+      </div>
+    </Dialog>
+  )
+}
