+++ web/components/properties/EditablePropertyCell.tsx
@@ -223,7 +223,6 @@ export function EditablePropertyCell({
           value={assigneeValue}
           allowTeams={true}
           allowUnassigned={true}
-          size="sm"
           onValueChanged={(next) => {
             const prev = assigneeValue
             if (next === prev) {
