+++ web/package.json
@@ -28,7 +28,7 @@
     "@dnd-kit/core": "6.3.1",
     "@dnd-kit/modifiers": "9.0.0",
     "@dnd-kit/sortable": "10.0.0",
-    "@helpwave/hightide": "^0.11.0",
+    "@helpwave/hightide": "^0.12.4",
     "@helpwave/internationalization": "0.4.0",
     "@tailwindcss/postcss": "4.3.1",
     "@tanstack/react-query": "5.101.0",
